:construction: Add dda raycasting not working yet
*

+++ inc/structs.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   structs.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: foctavia <foctavia@student.42.fr>          +#+  +:+       +#+        */
+/*   By: owalsh <owalsh@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/27 09:16:02 by owalsh            #+#    #+#             */
-/*   Updated: 2022/11/24 11:53:26 by foctavia         ###   ########.fr       */
+/*   Updated: 2022/11/28 17:33:45 by owalsh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,8 +59,8 @@ typedef struct s_checker
 
 typedef struct	s_coord
 {
-	float	x;
-	float	y;	
+	double	x;
+	double	y;	
 }				t_coord;
 
 typedef struct s_map
